Added a small helper program.
A real hack of a program! But it kind of works at least...

Let's you feed it a variable declaration, for example, feed it "int *a" and it ouputs:

```
int *a;
size_t a_count;
cargo_add_option(cargo, "--a", "Description of a", "[i]+", &a, &a_count); // Allocated unlimited length.
cargo_add_option(cargo, "--a", "Description of a", "[i]#", &a, &a_count, 128); // Allocated with max length 128.
```