bugfix: Make sure we always get a valid max width value.
Make sure we use the correct types, and assign a valid value in all code paths.

cargo_get_console_width returns an int, and returns -1 on error. So doing:

```
size_t console_width = cargo_get_console_width();
```

is not good, since on an error in cargo_get_console_width, the console_width will be parsed as the MAX value to fit in a `size_t`. So my check for `> 0` will always result in the true branch hitting.