Bool counting is now specified in the format.
The option flags should not keep stuff that fit better in the format string.

"b!" means that hte flag can be repeated, and each occurance will be counted, including -vvvv
"b=" means that the user specifies the flag target, followed by an integer that will be stored into the boolean target.