Check for unknown options after parsing by default.
Before it was done before parsing. Which kind of made CARGO_OPT_STOP useless unless you turn off to fail on unknown options, which sucks.

Because if you have multiple parsers, and the first parser does not know any of the variables after a certain option is specified. If we check for unknown options before we know this stop point we will fail the parse and never be able to pass on the remaining args to the second parse that might know about them.